requirements.txt
google-cloud-storage>=2.7.0
fastapi>=0.68.0
uvicorn>=0.15.0
torch>=2.0.0
transformers>=4.30.0
rank_bm25>=0.2.1
nltk>=3.6.2
google-auth>=2.0.0
